Revert "Add process input for options menu"
This reverts commit b6c8fb10d31f99cf9d7fcf8dae4045fae2954d80.

+++ src/MenuState.cpp
@@ -99,7 +99,7 @@ void MenuState::update(float delta){
 		options[current_option]->set_pos(FONT_X, FONT_Y, true, true);
 		options[current_option]->set_color(LIGHT_GREEN);
 
-		for(int idx = 0; idx < (int)options.size(); idx++){
+		for(int idx = 0; idx < options.size(); idx++){
 			if(idx == current_option) continue;
 			bool before = idx < current_option;
 			Text* option = options[idx + (before ? 1 : -1)];
